requirements.txt
Django==1.5.1
PIL==1.1.7
South==0.8.1
argparse==1.2.1
cmsplugin-contact==1.0.0
dj-database-url==0.2.2
dj-static==0.0.5
django-classy-tags==0.4
django-cms==2.4.3
django-form-designer==0.7.0
django-mptt==0.5.2
django-picklefield==0.3.1
django-sekizai==0.7
django-toolbelt==0.0.1
djangorestframework==2.3.12
gunicorn==18.0
heroku==0.1.4
html5lib==1.0b1
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==1.5
recaptcha-client==1.0.6
requests==2.2.1
six==1.3.0
static==1.0.2
wsgiref==0.1.2
